backend/requirements.txt
pdf2image
pillow
litellm==1.65.4
fastapi[standard]==0.112.0
pymupdf==1.22.5
llama-index==0.12.31
PyPDF2